UDP Data example

+++ server/src/com/swarmer/server/game/aco/graph/Vector2.java
@@ -2,10 +2,10 @@ package com.swarmer.server.game.aco.graph;
 
 public class Vector2 {
 
-    public float x;
-    public float y;
+    public int x;
+    public int y;
 
-    public Vector2(float x, float y) {
+    public Vector2(int x, int y) {
         this.x = x;
         this.y = y;
     }
@@ -15,13 +15,13 @@ public class Vector2 {
         this.y = 0;
     }
 
-    public void setLength(float size) {
+    public void setLength(int size) {
         x = x / getLength() * size;
         y = y / getLength() * size;
     }
 
-    public float getLength() {
-        return (float) Math.sqrt(Math.pow(x, 2) + Math.pow(y, 2));
+    public int getLength() {
+        return (int) Math.sqrt(Math.pow(x, 2) + Math.pow(y, 2));
     }
 
     @Override public String toString() {
